refractor: reorganize components folder
