+++ src/components/ClipControls/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ClipControls.jsx";
