+++ src/components/ColorMapControls/ColorMapControls.jsx
@@ -10,7 +10,7 @@ import {
 } from "@reach/listbox";
 import "@reach/listbox/styles.css";
 
-import Section from "./Section.jsx";
+import Section from "../Section/Section.jsx";
 
 function ColorMapControls({ colorMaps, colorMap, setColorMap }) {
   return (
